split api files based on user facing and internal
